% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/company_hierarchy.R
\name{add_company_hierarchy_stats}
\alias{add_company_hierarchy_stats}
\title{Add hierarchy stats to all the vertices of a company hierarchy graph}
\usage{
add_company_hierarchy_stats(graph)
}
\arguments{
\item{graph}{A graph}
}
\value{
A graph where all the nodes contain information about the company hierarchy graph
}
\description{
Add hierarchy stats to all the vertices of a company hierarchy graph
}
\keyword{company}
\keyword{graph}
\keyword{hierarchy}
