% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/company_hierarchy.R
\name{aggregate_company_hierarchy_value}
\alias{aggregate_company_hierarchy_value}
\title{Adds an aggregate value to the vertices of a company hierarchy}
\usage{
aggregate_company_hierarchy_value(graph, name_attribute, name_aggregate,
  FUN, ...)
}
\arguments{
\item{graph}{A graph}

\item{name_attribute}{The name of the value attribute to be aggregated}

\item{name_aggregate}{The name of the attribute where the aggregated value is stored}

\item{FUN}{the function which is used to calculate aggregated}

\item{...}{The parameters passed to the function specified in FUN}
}
\value{
A graph where all the nodes contain aggregated value
}
\description{
Adds an aggregate value to the vertices of a company hierarchy
}
\keyword{company}
\keyword{graph}
\keyword{hierarchy}
