% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anonymization.R
\name{anonymize_tbl}
\alias{anonymize_tbl}
\title{Generating a new 'anonymized' version of a dataframe}
\usage{
anonymize_tbl(tbl_in, vec_col_names, perc_dev = 0.2)
}
\arguments{
\item{tbl_in}{The data frame containing the data to be anonymized}

\item{vec_col_names}{vector of column names that need to be anonymized}

\item{lat}{The name of the column containing the latitude data}
}
\description{
The data frame will be anonymized by doing 3 things:
\enumerate{
  \item Replicating about \emph{italics}perc_dev of the rows randomly.
  \item Getting a random subset from those of about \emph{italics}1 - perc_dev
  \item Randomizing numeric values by multiplying them by random values between 1 - perc_dev and 1 + perc_dev
  \item Randomizing non-numeric values by reassigning the randomly to new rows.
}
}
\examples{
tbl_companies_uk_anon <- anonymize_tbl(tbl_companies_uk,
                                       vec_col_names = c("LONGITUDE_RA",
                                                         "LATITUDE_RA",
                                                         "ESTABLISHMENT_DATE"))
}
\keyword{anonymization}
\keyword{demo}
