% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_config.R
\name{apply_column_config}
\alias{apply_column_config}
\title{A function for applying column configuration file}
\usage{
apply_column_config(df_source, filename_source)
}
\arguments{
\item{df_source}{The dataframe of which you want to apply/get the import specs for}

\item{filename_source}{The filename where the data frame is imported fromset}
}
\value{
A data frame containing column importing specs
}
\description{
This function applies a column configuration for a source file which
you can use to control a data file import .
The first time it is used it will create a column specification file.
}
\examples{
get_columns_file(df, filename_source)
}
\keyword{config}
\keyword{data}
\keyword{import}
