% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_config.R
\name{auto_rename_columns}
\alias{auto_rename_columns}
\title{Autorenames columns based on word detection and adding a prefix}
\usage{
auto_rename_columns(column_names, detection_words, prefix)
}
\arguments{
\item{column_names}{a vector with column}

\item{detection_words}{The words that indicate a prefix should be added to the column name}

\item{prefix}{The prefix the should be added in case a word is detected}
}
\value{
A boolean indicating whether the two column specs are equal
}
\description{
Autorenames columns based on word detection and adding a prefix
}
\examples{
clean_column_names(column_names)
}
\keyword{config}
\keyword{data}
\keyword{import}
