% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_tools.R
\name{bin_quantity_labeled}
\alias{bin_quantity_labeled}
\title{Creates a vector of human readable bins based on width}
\usage{
bin_quantity_labeled(values, qty_bins, FUN, ...)
}
\arguments{
\item{values}{The vector/column of values you want to 'ntile'}

\item{qty_bins}{The number of ntiles you want to have}

\item{FUN}{The function you want to use to apply formatting}

\item{...}{The parameters you'd want to pass on to the formatting function supplied to FUN}
}
\value{
Ordered factor
}
\description{
Creates a vector of human readable bins based on width
}
\examples{
bin_quantity_labeled(values = mtcars$mpg, qty_ntiles = 4, FUN = format_currency, currency = "GBP")
}
