% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_config.R
\name{clean_column_names}
\alias{clean_column_names}
\title{Clean column names}
\usage{
clean_column_names(column_names)
}
\arguments{
\item{column_names}{a vector with column}
}
\value{
A boolean indicating whether the two column specs are equal
}
\description{
This function allows you to read data frame specs into a data frame
}
\examples{
clean_column_names(column_names)
}
\keyword{config}
\keyword{data}
\keyword{import}
