% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/economic_activity_hierachy.R
\name{clean_hierarchy}
\alias{clean_hierarchy}
\title{cleans up all nace codes from the hierarchy that contain a NA value and are non-connective}
\usage{
clean_hierarchy(tbl_hierarchy)
}
\arguments{
\item{tbl_hierarchy}{a data frame that should contain the entire NACE or SBI hierarchy with a
quantity or amount. It should contain the following columns: \describe{
  \item{code}{a NACE or SBI code.}
  \item{code_parent}{a NACE or SBI code that refers to the direct parent code.}
  \item{layer_no}{an integer indicating the hierarchy's level, 1 being the top level}
  \item{value}{the quantity that is used to evaluate whether the code should be kept in
  place or should be 'pushed up' the hierarchy.}
}}
}
\value{
a data frame containing only the codes with other values than NA and are non-connecting
}
\description{
cleans up all nace codes from the hierarchy that contain a NA value and are non-connective
}
\keyword{NACE}
\keyword{SBI}
\keyword{SIC}
