% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{col_graydon}
\alias{col_graydon}
\title{Set of Graydon company colors}
\format{An object of class \code{character} of length 9.}
\usage{
col_graydon
}
\description{
A vector containing 9 Graydon company colors in hexadecimal format.
}
\keyword{datasets}
