% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{col_graydon_high}
\alias{col_graydon_high}
\title{High value color}
\format{An object of class \code{character} of length 1.}
\usage{
col_graydon_high
}
\description{
A variable containing a Graydon company color used for the highest value in positive
valued gradients. The color value is in hexadecimal format.
}
\keyword{datasets}
