% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/company_hierarchy.R
\name{combine_graphs}
\alias{combine_graphs}
\title{Combining graphs from a list to a single graph}
\usage{
combine_graphs(list_graphs)
}
\arguments{
\item{list_graphs}{A list of graphs}
}
\value{
A graph containing all deduplicated graphs
}
\description{
Combining graphs from a list to a single graph
}
\examples{
graph_total <- combine_graphs(list_graphs)
}
\keyword{company}
\keyword{graph}
\keyword{hierarchy}
