% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uncategorized_tools.R
\name{create_gitignore}
\alias{create_gitignore}
\title{Creates standard gitignore file for a project. ----}
\usage{
create_gitignore()
}
\description{
Creates standard gitignore file for a project. ----
}
