% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geomap_tools.R
\name{create_spatial_df}
\alias{create_spatial_df}
\title{Make shape file of data frame from long and lat columns}
\usage{
create_spatial_df(df, lon = "LONGITUDE_RA", lat = "lat")
}
\arguments{
\item{df}{The data frame containing the data to be mapped}

\item{lon}{The name of the column containing the longitude data}

\item{lat}{The name of the column containing the latitude data}
}
\description{
Make shape file of data frame from long and lat columns
}
\examples{
sp_companies_uk <- create_spatial_df(tbl_companies_uk, lon = "LONGITUDE_RA", lat = "LATITUDE_RA")
}
\keyword{maps}
