% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_config.R
\name{determine_column_specs}
\alias{determine_column_specs}
\title{A function to determine column specs for a data frame}
\usage{
determine_column_specs(df_source)
}
\arguments{
\item{df_source}{The dataframe of which you want to read the column specs from}
}
\value{
A data frame containing the column specs of df_source
}
\description{
This function allows you to read data frame specs into a data frame
}
\examples{
determine_column_specs(df)
}
\keyword{config}
\keyword{data}
\keyword{import}
