% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uncategorized_tools.R
\name{df_to_clipboard}
\alias{df_to_clipboard}
\title{Copy a dataframe to the clipboard.}
\usage{
df_to_clipboard(df, row.names = FALSE, col.names = TRUE, ...)
}
\arguments{
\item{df}{The data frame you want to put on the clipboard}

\item{row.names}{Indicates whether you want to copy the row names as well, default is FALSE.}

\item{col.names}{Indicates whether you want to copy the column names as well, default is TRUE.}
}
\description{
Copy a dataframe to the clipboard.
}
\examples{
df_to_clipboard(mtcars)
}
