% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geomap_tools.R
\name{distance_degrees_to_km}
\alias{distance_degrees_to_km}
\title{Formula to calculate distances between two lon-lat points}
\usage{
distance_degrees_to_km(lon1, lat1, lon2, lat2)
}
\arguments{
\item{lon1}{The name of the column containing the first point longitude}

\item{lat1}{The name of the column containing the first point latitude}

\item{lon2}{The name of the column containing the second point longitude}

\item{lat2}{The name of the column containing the second point latitude}
}
\description{
Formula to calculate distances between two lon-lat points
}
\examples{
kilometers <- distance_degrees_to_km(lon1, lat1, lon2, lat2)
}
\keyword{distance}
\keyword{latitude}
\keyword{longitude}
\keyword{maps}
