% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uncategorized_tools.R
\name{end_of_month}
\alias{end_of_month}
\title{Gets the last day of the month of a date in the format YYYY-MM-DD}
\usage{
end_of_month(date, tz = "CET")
}
\arguments{
\item{date}{you'll want to get the last month}

\item{tz}{The time zone indicator, which has a default for the Central European time zone}
}
\description{
Gets the last day of the month of a date in the format YYYY-MM-DD
}
\examples{
end_of_month(as.Date('2018-10-09'), tz = "GMT")
}
