% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_tools.R
\name{format_currency}
\alias{format_currency}
\title{Formats a number into a currency string}
\usage{
format_currency(amount, currency = c("EUR", "GBP"),
  number_decimals = 2, scale = NA)
}
\arguments{
\item{amount}{The number you want to format as euros}

\item{currency}{The currency symbol that should be used as a prefix}

\item{number_decimals}{The largest number that should still be formatted using cents}

\item{scale}{Used when you want to format the number in scales (regular, per thousands or millions)}
}
\value{
A string containing the euro formatted number
}
\description{
Formats a number into a currency string
}
\examples{
format_currency(amount = 1000000, currency = "EUR", number_decimals = 1,scale = "M")
}
