% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_tools.R
\name{format_number}
\alias{format_number}
\title{Sets first letter of a string to capital}
\usage{
format_number(x, number_decimals = 0, format_EN = FALSE,
  scale = c("normal", "k", "M"))
}
\arguments{
\item{x}{The string you want to apply the formatting to}

\item{number_decimals}{The number of decimals you want to display (default = 0)}

\item{format_EN}{use EN notation; meaning '.' a decimal and ',' as thousands seperators.}

\item{scale}{Used when you want to format the amount in scales (regular, per thousands or millions)}
}
\value{
A formatted string
}
\description{
Sets first letter of a string to capital
}
\examples{
format_number(1000000.01)
}
