% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_config.R
\name{get_columns_file}
\alias{get_columns_file}
\title{A function for creating/reading data source column specs}
\usage{
get_columns_file(df_source, filename_source)
}
\arguments{
\item{df_source}{The dataframe of which you want to apply/get the import specs for}

\item{filename_source}{The filename where the data frame is imported fromset}
}
\value{
A data frame containing column importing specs
}
\description{
This function allows you to create and or read a file which you can use to control a data file import
}
\examples{
get_columns_file(df, filename_source)
}
\keyword{config}
\keyword{data}
\keyword{import}
