% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/economic_activity_hierachy.R
\name{get_incoming_vertice_names}
\alias{get_incoming_vertice_names}
\title{Get the names of the vertices that are children of the given vertex}
\usage{
get_incoming_vertice_names(graph, name_vertex, order = 1)
}
\arguments{
\item{graph}{The graph to do the searhing in}

\item{name_vertex}{The name of the vertex to get the children from}

\item{order}{The number of connections to search through to get the children}
}
\value{
List of vertex names that are incoming
}
\description{
Get the names of the vertices that are children of the given vertex
}
\examples{
get_incoming_vertice_names(graph = graph_SBI, name_vertex = "42", order = 1)
}
\keyword{NACE}
\keyword{SBI}
\keyword{SIC}
