% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/company_hierarchy.R
\name{get_qty_siblings}
\alias{get_qty_siblings}
\title{Get the number of sibling companies}
\usage{
get_qty_siblings(graph, vertx)
}
\arguments{
\item{graph}{A graph}

\item{vertx}{The vertex representing a company}
}
\value{
A data-frame containing all sibling companies
}
\description{
Get the number of sibling companies
}
\keyword{company}
\keyword{graph}
\keyword{hierarchy}
