% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/company_hierarchy.R
\name{get_root_vertex_name}
\alias{get_root_vertex_name}
\title{Get the root node's name of a tree}
\usage{
get_root_vertex_name(tree_graph)
}
\arguments{
\item{tree_graph}{The graph containing the hierarchical tree}
}
\value{
The vertex name that is the root of the graph
}
\description{
Get the root node's name of a tree
}
\examples{
graph_SBI <- get_root_vertex_name(tree_graph)
}
\keyword{company}
\keyword{graph}
\keyword{hierarchy}
