% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/company_hierarchy.R
\name{get_siblings_df}
\alias{get_siblings_df}
\title{Add a logical variable to the companies in a hierarchy when one of it's
attributes match one of a set of values}
\usage{
get_siblings_df(graph, company_ids)
}
\arguments{
\item{graph}{A graph}

\item{company_ids}{The name of the newly created logical company attribute}
}
\value{
A data-frame containing all sibling companies
}
\description{
Add a logical variable to the companies in a hierarchy when one of it's
attributes match one of a set of values
}
\keyword{company}
\keyword{graph}
\keyword{hierarchy}
