% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/economic_activity_hierachy.R
\name{graph_remove_empty_non_connecting}
\alias{graph_remove_empty_non_connecting}
\title{cleans up all nace codes from the hierarchy that contain a NA value and are non-connective}
\usage{
graph_remove_empty_non_connecting(graph_tree, name_attribute)
}
\arguments{
\item{graph_tree}{The graph containing the hierarchical tree}

\item{name_attribute}{name of the attribute you want to evaluate to be non-zero}
}
\value{
A graph containing only the codes with other values than NA and are non-connecting
}
\description{
cleans up all nace codes from the hierarchy that contain a NA value and are non-connective
}
\examples{
graph_remove_empty_non_connecting(graph_SBI, name_attribute = "qty_companies")
}
\keyword{NACE}
\keyword{SBI}
\keyword{SIC}
