% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graydon.package-package.r
\docType{package}
\name{graydon.package}
\alias{graydon.package}
\alias{graydon.package-package}
\title{graydon.package: toolkit for faster development}
\description{
graydon.package: toolkit for faster development
}
\section{ggplot2 functions}{

theme_graydon - A function for applying a Graydon theme to a ggplot
}

