% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_config.R
\name{has_equal_column_specs}
\alias{has_equal_column_specs}
\title{Test whether two column specs are equal}
\usage{
has_equal_column_specs(tbl_column_specs_a, tbl_column_specs_b)
}
\arguments{
\item{tbl_column_specs_a}{The dataframe with column specs}

\item{tbl_column_specs_b}{The dataframe with column specs}
}
\value{
A boolean indicating whether the two column specs are equal
}
\description{
This function allows you to read data frame specs into a data frame
}
\examples{
has_equal_column_specs(df_specs_a, df_specs_b)
}
\keyword{config}
\keyword{data}
\keyword{import}
