% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/company_hierarchy.R
\name{hierarchy_as_data_frame}
\alias{hierarchy_as_data_frame}
\title{Converts a company hierarchy graph to a data frame}
\usage{
hierarchy_as_data_frame(graph)
}
\arguments{
\item{list_graphs}{The graph containing all the hierarchical company trees}
}
\value{
A data frame with company hierarchy data
}
\description{
Converts a company hierarchy graph to a data frame
}
\keyword{company}
\keyword{graph}
\keyword{hierarchy}
