% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/economic_activity_hierachy.R
\name{hierarchy_code_all}
\alias{hierarchy_code_all}
\title{Enrich each of the codes with all of the codes from the levels above the current code}
\usage{
hierarchy_code_all(tbl_hierarchy)
}
\arguments{
\item{tbl_hierarchy}{a data frame that should contain the entire NACE or SBI hierarchy with a
quantity or amount. It should contain the following columns: \describe{
  \item{code}{a NACE or SBI code.}
  \item{code_parent}{a NACE or SBI code that refers to the direct parent code.}
  \item{layer_no}{an integer indicating the hierarchy's level, 1 being the top level}
}}
}
\value{
a data frame containing the complete tbl_hierarchy data frame, adding the column ode_level_x (the code at the chosen level of the hierarchy)
}
\description{
Enrich each of the codes with all of the codes from the levels above the current code
}
\keyword{NACE}
\keyword{SBI}
\keyword{SIC}
