% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/economic_activity_hierachy.R
\name{hierarchy_code_level}
\alias{hierarchy_code_level}
\title{A function for enriching each of the codes from it's chosen level's code value}
\usage{
hierarchy_code_level(tbl_hierarchy, level_no)
}
\arguments{
\item{tbl_hierarchy}{a data frame that should contain the entire NACE or SBI hierarchy with a
quantity or amount. It should contain the following columns: \describe{
  \item{code}{a NACE or SBI code.}
  \item{code_parent}{a NACE or SBI code that refers to the direct parent code.}
  \item{layer_no}{an integer indicating the hierarchy's level, 1 being the top level}
  \item{value}{the quantity that is used to evaluate whether the code should be kept in
  place or should be 'pushed up' the hierarchy.}
}}

\item{tbl_hierarchy}{a data frame that should contain the entire NACE or SBI hierarchy with
the added code from the requested level: \describe{
  \item{code}{a NACE or SBI code.}
  \item{code_parent}{a NACE or SBI code that refers to the direct parent code.}
  \item{layer_no}{an integer indicating the hierarchy's level, 1 being the top level}
  \item{code_level_x}{The code at the chosen level of the hierarchy}
}}
}
\description{
A function for enriching each of the codes from it's chosen level's code value
}
\examples{
hierarchy_code_level(tbl_hierarchy, level_no = 2)
}
\keyword{NACE}
\keyword{SBI}
\keyword{SIC}
