% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/company_hierarchy.R
\name{hierarchy_list_as_data_frame}
\alias{hierarchy_list_as_data_frame}
\title{Converts a list of company hierarchy graphs to a data frame}
\usage{
hierarchy_list_as_data_frame(list_graphs)
}
\arguments{
\item{list_graphs}{The graph containing all the hierarchical company trees}
}
\value{
A data frame with company hierarchy data
}
\description{
Converts a list of company hierarchy graphs to a data frame
}
\examples{
df_hierarchies <- hierarchy_list_as_data_frame(list_graph_hierarchies)
}
\keyword{company}
\keyword{graph}
\keyword{hierarchy}
