% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uncategorized_tools.R
\name{install_graydon_packages}
\alias{install_graydon_packages}
\title{Installing regularly used packages}
\usage{
install_graydon_packages()
}
\description{
Installing regularly used packages
}
\examples{
install_graydon_packages()
}
