% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_tools.R
\name{label_intervals}
\alias{label_intervals}
\title{Creates a vector of human readable bins}
\usage{
label_intervals(values, intervals, FUN, ...)
}
\arguments{
\item{values}{The original values}

\item{intervals}{The created intervals}

\item{FUN}{The function you want to use to apply formatting}

\item{...}{The parameters you'd want to pass on to the formatting function supplied to FUN}
}
\value{
Ordered factor
}
\description{
Creates a vector of human readable bins
}
\examples{
label_intervals(vec_values = mtcars$mpg, intervals, FUN = format_currency, currency = "GBP")
}
