% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/company_hierarchy.R
\name{mark_companies_logical}
\alias{mark_companies_logical}
\title{Add a logical variable to the companies in a hierarchy when one of it's
attributes match one of a set of values}
\usage{
mark_companies_logical(graph, name_logical, name_filter, set_criteria)
}
\arguments{
\item{graph}{A graph}

\item{name_logical}{The name of the newly created logical company attribute}

\item{name_filter}{The name of the attribute which values are compared to the criteria}

\item{set_criteria}{The set of criteria in a vector to which the values are compared}
}
\value{
A graph where all the nodes contain the newly created attribute
}
\description{
Add a logical variable to the companies in a hierarchy when one of it's
attributes match one of a set of values
}
\keyword{company}
\keyword{graph}
\keyword{hierarchy}
