% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/company_hierarchy.R
\name{mark_ego_graph}
\alias{mark_ego_graph}
\title{Function earmark the companies in the neighbourhood of a company, this neighbourhood is sometimes
called [ego graphs](http://mathworld.wolfram.com/NeighborhoodGraph.html)}
\usage{
mark_ego_graph(graph, id_company, target_attribute, distance = 1,
  direction = "in")
}
\arguments{
\item{graph}{A graph containing company/company relations data.}

\item{id_company}{A company id of which you want to mark the neighbourhood}

\item{target_attribute}{The name of the new attribute you want to mark the neighborhood in}

\item{distance}{The number of 'hops' in the company network that should be included, default = 1}

\item{direction}{The direction to calculate the propagated value from (can be "in", "out", "all")}
}
\value{
A graph containing the newly added neighbourhood marking attribute
}
\description{
Function earmark the companies in the neighbourhood of a company, this neighbourhood is sometimes
called [ego graphs](http://mathworld.wolfram.com/NeighborhoodGraph.html)
}
\examples{
graph_all_companies <- mark_ego_graph(graph_all_companies, id_company = "910716048", target_attribute = "is_neighbour")
}
\keyword{company}
\keyword{ego}
\keyword{graph}
\keyword{hierarchy}
