% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geomap_tools.R
\name{match_sp}
\alias{match_sp}
\title{Find key for merging Spatial* objects}
\usage{
match_sp(sp_from, sp_to)
}
\arguments{
\item{df}{The data frame containing the data to be mapped}

\item{lon}{The name of the column containing the longitude data}

\item{lat}{The name of the column containing the latitude data}
}
\description{
Find key for merging Spatial* objects
}
\examples{
match_sp(sp_companies_uk, sp_uk_lieutenancy)
}
\keyword{maps}
