% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_tools.R
\name{ntiles_labeled}
\alias{ntiles_labeled}
\title{Creates a vector of human readable ntiles}
\usage{
ntiles_labeled(values, qty_ntiles, use_intervals = FALSE,
  format_EN = FALSE, FUN, ...)
}
\arguments{
\item{qty_ntiles}{The number of ntiles you want to have}

\item{FUN}{The function you want to use to apply formatting}

\item{...}{The parameters you'd want to pass on to the formatting function supplied to FUN}

\item{vec_values}{The vector/column of values you want to 'ntile'}
}
\value{
Ordered factor
}
\description{
Creates a vector of human readable ntiles
}
\examples{
ntiles_labeled(vec_values = mtcars$mpg, qty_ntiles = 4, FUN = format_currency, currency = "GBP")
}
