% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/economic_activity_hierachy.R
\name{plot_econ_hierarchy}
\alias{plot_econ_hierarchy}
\title{Plotting hierarchies based on a single value, mostly for example puroposes}
\usage{
plot_econ_hierarchy(graph, title = "", label = NA, size = NA, ...)
}
\arguments{
\item{graph}{The graph made of a economic activity tree (usually created with the create_economic_activity_graph function)}

\item{title}{The title you want displayed in the plot}

\item{label}{The name of the vertex attribute you want to use as a label}

\item{size}{The name of the vertex attribute you want to use for size}
}
\description{
Plotting hierarchies based on a single value, mostly for example puroposes
}
\examples{
plot_econ_hierarchy(graph_SBI)
}
\keyword{NACE}
\keyword{SBI}
\keyword{SIC}
