% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/company_hierarchy.R
\name{plot_graydon_graph}
\alias{plot_graydon_graph}
\title{Plots a company hierarchy graph}
\usage{
plot_graydon_graph(graph, ...)
}
\arguments{
\item{graph}{A graph}

\item{...}{The parameters passed to the function specified in FUN}
}
\description{
Plots a company hierarchy graph
}
\examples{
plot_graydon_graph(graph)
}
\keyword{company}
\keyword{graph}
\keyword{hierarchy}
