% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/economic_activity_hierachy.R
\name{plot_hierarchy}
\alias{plot_hierarchy}
\title{Plotting hierarchies based on a single value, mostly for example puroposes}
\usage{
plot_hierarchy(tbl_hierarchy, title = "")
}
\arguments{
\item{tbl_hierarchy}{a data frame that should contain the entire NACE or SBI hierarchy with a
quantity or amount. It should contain the following columns: \describe{
  \item{code}{a NACE or SBI code.}
  \item{code_parent}{a NACE or SBI code that refers to the direct parent code.}
  \item{layer_no}{an integer indicating the hierarchy's level, 1 being the top level}
  \item{value}{the quantity that is used to evaluate whether the code should be kept in
  place or should be 'pushed up' the hierarchy.}
}}
}
\description{
Plotting hierarchies based on a single value, mostly for example puroposes
}
\keyword{NACE}
\keyword{SBI}
\keyword{SIC}
