% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/company_hierarchy.R
\name{propagate_hierarchy_value}
\alias{propagate_hierarchy_value}
\title{Adds an aggregate value to the vertices of a company hierarchy, in which you can vary it's direction and distance}
\usage{
propagate_hierarchy_value(graph, name_attribute, name_propagate,
  distance = 1, direction = "in", FUN, ...)
}
\arguments{
\item{graph}{A graph}

\item{name_attribute}{The name of the value attribute to be aggregated}

\item{name_propagate}{The name of the attribute where the aggregated value is stored}

\item{distance}{The number of 'hops' in the company network that should be included, default = 1}

\item{direction}{The direction to calculate the propagated value from (can be "in", "out", "all")}

\item{FUN}{the function which is used to calculate aggregated}

\item{...}{The parameters passed to the function specified in FUN}
}
\value{
A graph where all the nodes contain aggregated value
}
\description{
Adds an aggregate value to the vertices of a company hierarchy, in which you can vary it's direction and distance
}
\examples{
graph <- propagate_hierarchy_value(graph, name_attribute = "qty_employees", name_propagate = "qty_employees_cum", distance = 1, direction = "in", FUN = sum, na.rm = TRUE)
}
\keyword{company}
\keyword{graph}
\keyword{hierarchy}
