% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anonymization.R
\name{randomize_num_column}
\alias{randomize_num_column}
\title{Anonymize the values of numeric columns by multiplying their current values}
\usage{
randomize_num_column(column, perc_dev)
}
\description{
Anonymize the values of numeric columns by multiplying their current values
}
