% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_config.R
\name{read_colspecs_file}
\alias{read_colspecs_file}
\title{Read column specs from a file}
\usage{
read_colspecs_file(filename_columns)
}
\arguments{
\item{filename_columns}{The file name containing the column specifications}
}
\value{
A data frame containing the column specifications
}
\description{
Read column specs from a file
}
\examples{
tbl_column_specs <- read_colspecs_file(filename_specs)
}
\keyword{config}
\keyword{data}
\keyword{import}
