% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/company_hierarchy.R
\name{recode_holding_codes}
\alias{recode_holding_codes}
\title{Determine holding SBI replacement}
\usage{
recode_holding_codes(graph_company_hierarchy, name_activity_code,
  vec_holding_codes)
}
\arguments{
\item{graph_company_hierarchy}{A graph with the company hierarchy}

\item{name_activity_code}{The name of the value attribute that contains the economic activity code}

\item{vec_holding_codes}{A vector of codes that represent holdings}
}
\value{
A graph where all the nodes contain aggregated value
}
\description{
Determine holding SBI replacement
}
\examples{
graph_company_hierarchy <- recode_holding_codes(graph_company_hierarchy, name_activity_code = "code_sbi", vec_holding_codes = c("64", "642", "6420"))
}
\keyword{company}
\keyword{graph}
\keyword{hierarchy}
