% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uncategorized_tools.R
\name{remove_redundant_columns}
\alias{remove_redundant_columns}
\title{Determining redundant columns of secondary table before joining}
\usage{
remove_redundant_columns(tbl_primary, tbl_secondary, vec_key_columns)
}
\arguments{
\item{tbl_primary}{The table of which you}

\item{tbl_secondary}{The time zone indicator, which has a default for the Central European time zone}

\item{vec_key_columns}{a vector indicating the column which are the common keys in both tables}
}
\description{
Determining redundant columns of secondary table before joining
}
\examples{
remove_redundant_columns(tbl_primary, tbl_secondary, vec_key_columns)
}
