% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_tools.R
\name{retrieve_sp_map}
\alias{retrieve_sp_map}
\title{A function to create a SpatialPolygonsDataFrame}
\usage{
retrieve_sp_map(map_directory, country, layer)
}
\arguments{
\item{map_directory}{The directory where the shap files will be stored}

\item{map_directory}{The type of grid or background}
}
\description{
The first time it is used it will create a column specification file
allows you to create and or read a file which you can use to control a data file import
}
\examples{
theme_graydon("grid")
}
\keyword{maps}
