% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/economic_activity_hierachy.R
\name{roll_up_hierarchy}
\alias{roll_up_hierarchy}
\title{A function to determine which economic activity definitiion level is optimal if you want to have a minimum value in a economic activity group}
\usage{
roll_up_hierarchy(tbl_hierarchy, threshold)
}
\arguments{
\item{tbl_hierarchy}{a data frame that should contain the entire NACE or SBI hierarchy with a
quantity or amount. It should contain the following columns: \describe{
  \item{code}{a NACE or SBI code.}
  \item{code_parent}{a NACE or SBI code that refers to the direct parent code.}
  \item{layer_no}{an integer indicating the hierarchy's level, 1 being the top level}
  \item{value}{the quantity that is used to evaluate whether the code should be kept in
  place or should be 'pushed up' the hierarchy.}
}}

\item{threshold}{The minimum value that the tbl_hierarchy value column should have to leave a
code in it's place.}
}
\value{
a data frame containing \describe{
  \item{code}{The original code which was input.}
  \item{code_new}{The replacement code, which indicates the new position in the hierarchy}
  \item{value}{The value associated with the original code (input)}
}
}
\description{
This function can be used to aggregate a NACE or SBI code economic activity tree,
so the codes represent enough of something for example number of companies,
number of customers or revenue.
}
\examples{
roll_up_hierarchy(tbl_hierarchy, threshold = 200)
}
\keyword{NACE}
\keyword{SBI}
\keyword{SIC}
