% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/economic_activity_hierachy.R
\name{roll_up_hierarchy_by_level}
\alias{roll_up_hierarchy_by_level}
\title{Roll up economic activity hierarchy to a certain level in the tree}
\usage{
roll_up_hierarchy_by_level(graph_tree, level_to, name_attribute)
}
\arguments{
\item{graph_tree}{Graph representing the economic activity hierarchy}

\item{level_to}{The level to which you want to move the nodes in the hierarchy}

\item{name_attribute}{The name of the attribute you want to check for values}
}
\value{
A data frame containing the original economic activity code, the new activity code and the quantity/value that the new code would contain if aggregated
}
\description{
This function can be used to aggregate a NACE or SBI code economic activity tree,
so the codes represent enough of something for example number of companies, number of customers or
revenue.
}
\examples{
graph_SBI_rolled <- roll_up_hierarchy_by_level(graph_tree = graph_SBI,
                                               level_to = 2
                                               name_attribute = "qty_companies")
}
\keyword{NACE}
\keyword{SBI}
\keyword{SIC}
