% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/economic_activity_hierachy.R
\name{rolled_up_as_data_frame}
\alias{rolled_up_as_data_frame}
\title{Converts a company hierarchy graph to a data frame}
\usage{
rolled_up_as_data_frame(graph)
}
\arguments{
\item{graph}{The graph containing all the hierarchical company trees}
}
\value{
A data frame with company hierarchy data
}
\description{
Converts a company hierarchy graph to a data frame
}
\examples{
df_hierarchy <- hierarchy_as_data_frame(graph_hierarchy)
}
\keyword{company}
\keyword{graph}
\keyword{hierarchy}
