% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_tools.R
\name{save_plot_to_png}
\alias{save_plot_to_png}
\title{Saving a plot to a png, ready for use in a PowerPoint}
\usage{
save_plot_to_png(plot, file_name, squared = FALSE)
}
\arguments{
\item{plot}{the ggplot, stored as a variable, which you want to save}

\item{file_name}{the filename, excluding the extension, you ant to save the plot to}

\item{squared}{logical indicating whether the plot is squared or landscape,
conform to side by side side slide (squared = TRUE) or
One graph in a slide (squared = FALSE)}
}
\description{
Saving a plot to a png, ready for use in a PowerPoint
}
\keyword{ggplot2}
