% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_tools.R
\name{scale_color_graydon}
\alias{scale_color_graydon}
\title{A function to include in a ggplot so the col aesthetics make use of the Graydon color palette}
\usage{
scale_color_graydon()
}
\description{
A function to include in a ggplot so the col aesthetics make use of the Graydon color palette
}
\keyword{ggplot2}
