% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_tools.R
\name{scale_fill_graydon}
\alias{scale_fill_graydon}
\title{A function to include in a ggplot so the fill aesthetics make use of the Graydon color palette}
\usage{
scale_fill_graydon()
}
\description{
A function to include in a ggplot so the fill aesthetics make use of the Graydon color palette
}
\keyword{ggplot2}
