% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_tools.R
\name{scale_gradient_graydon}
\alias{scale_gradient_graydon}
\title{A function to include in a ggplot so color and fill aesthetics make use of the Graydon color palette}
\usage{
scale_gradient_graydon()
}
\description{
A function to include in a ggplot so color and fill aesthetics make use of the Graydon color palette
}
\keyword{ggplot2}
