% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_tools.R
\name{scale_x_currency}
\alias{scale_x_currency}
\title{A function to create a continuous x-axis for formatted currency}
\usage{
scale_x_currency(currency = c("EUR", "GBP"), number_decimals = 2,
  scale = c("normal", "k", "M"), ...)
}
\description{
A function to create a continuous x-axis for formatted currency
}
\keyword{ggplot2}
