% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_tools.R
\name{scale_x_numeric}
\alias{scale_x_numeric}
\title{A function to create a continuous x-axis for formatted numbers}
\usage{
scale_x_numeric(number_decimals = 0, format_EN = FALSE,
  scale = c("normal", "k", "M"), ...)
}
\description{
A function to create a continuous x-axis for formatted numbers
}
\keyword{ggplot2}
