% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_tools.R
\name{scale_x_percent}
\alias{scale_x_percent}
\title{A function to create a continuous x-axis for formatted percentages}
\usage{
scale_x_percent(number_decimals = 1, format_EN = FALSE, ...)
}
\description{
A function to create a continuous x-axis for formatted percentages
}
\keyword{ggplot2}
