% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sp_be_country}
\alias{sp_be_country}
\title{Map outlining the Belgium country borders}
\format{An object of class \code{SpatialPolygonsDataFrame} with 1 rows and 67 columns.}
\usage{
sp_be_country
}
\description{
\url{https://en.wikipedia.org/wiki/Belgium}
}
\keyword{datasets}
