% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sp_be_gewest}
\alias{sp_be_gewest}
\title{Map outlining the Belgium gewest borders}
\format{An object of class \code{SpatialPolygonsDataFrame} with 4 rows and 12 columns.}
\usage{
sp_be_gewest
}
\description{
\url{https://en.wikipedia.org/wiki/Communities,_regions_and_language_areas_of_Belgium}
}
\keyword{datasets}
