% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_be_postcode.R
\docType{data}
\name{sp_be_postcode}
\alias{sp_be_postcode}
\title{Map outlining the Belgian postal code borders}
\format{An object of class SpatialPolygonsDataFrame with 1268 rows and 6 columns.}
\source{
Shapefile data online available \href{https://www.geo.be/catalog/details/9738c7c0-5255-11ea-8895-34e12d0f0423?l=nl}{https://www.geo.be/catalog/details/9738c7c0-5255-11ea-8895-34e12d0f0423?l=nl}
}
\usage{
sp_be_postcode
}
\description{
Read more on Belgian postal codes online. See \href{https://en.wikipedia.org/wiki/Postal_codes_in_Belgium}{wikipedia}
}
\keyword{datasets}
