% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sp_be_province}
\alias{sp_be_province}
\title{Map outlining the Belgium province borders}
\format{An object of class \code{SpatialPolygonsDataFrame} with 192 rows and 14 columns.}
\usage{
sp_be_province
}
\description{
\url{https://en.wikipedia.org/wiki/Provinces_of_Belgium}
}
\keyword{datasets}
