% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sp_nl_corop}
\alias{sp_nl_corop}
\title{Map outlining the Dutch COROP borders}
\format{An object of class \code{SpatialPolygonsDataFrame} with 40 rows and 4 columns.}
\usage{
sp_nl_corop
}
\description{
\url{https://en.wikipedia.org/wiki/COROP}
}
\keyword{datasets}
