% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sp_nl_country}
\alias{sp_nl_country}
\title{Map outlining the Dutch country borders}
\format{An object of class \code{SpatialPolygonsDataFrame} with 1 rows and 3 columns.}
\usage{
sp_nl_country
}
\description{
\url{https://en.wikipedia.org/wiki/Netherlands}
}
\keyword{datasets}
