% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sp_nl_gemeenten}
\alias{sp_nl_gemeenten}
\title{Map outlining the Dutch manucipality borders}
\format{An object of class \code{SpatialPolygonsDataFrame} with 380 rows and 3 columns.}
\usage{
sp_nl_gemeenten
}
\description{
Map outlining the Dutch manucipality borders
}
\keyword{datasets}
