% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sp_nl_postcode}
\alias{sp_nl_postcode}
\title{Map outlining the Dutch postal code borders}
\format{An object of class \code{SpatialPolygonsDataFrame} with 4053 rows and 5 columns.}
\usage{
sp_nl_postcode
}
\description{
\url{https://en.wikipedia.org/wiki/Postal_codes_in_the_Netherlands}
}
\keyword{datasets}
