% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sp_nl_province}
\alias{sp_nl_province}
\title{Map outlining the Dutch province borders}
\format{An object of class \code{SpatialPolygonsDataFrame} with 12 rows and 2 columns.}
\usage{
sp_nl_province
}
\description{
\url{https://en.wikipedia.org/wiki/Provinces_of_the_Netherlands}
}
\keyword{datasets}
