% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sp_uk_country}
\alias{sp_uk_country}
\title{Map outlining the UK country borders}
\format{An object of class \code{SpatialPolygonsDataFrame} with 4 rows and 13 columns.}
\usage{
sp_uk_country
}
\description{
\url{https://en.wikipedia.org/wiki/Countries_of_the_United_Kingdom}
}
\keyword{datasets}
