% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sp_uk_lieutenancy}
\alias{sp_uk_lieutenancy}
\title{Map outlining the UK lieutenancy area borders}
\format{An object of class \code{SpatialPolygonsDataFrame} with 192 rows and 15 columns.}
\usage{
sp_uk_lieutenancy
}
\description{
\url{https://en.wikipedia.org/wiki/Lieutenancy_area}
}
\keyword{datasets}
