% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sp_uk_state}
\alias{sp_uk_state}
\title{Map outlining the UK state borders}
\format{An object of class \code{SpatialPolygonsDataFrame} with 1 rows and 68 columns.}
\usage{
sp_uk_state
}
\description{
\url{https://en.wikipedia.org/wiki/United_Kingdom}
}
\keyword{datasets}
