% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_tools.R
\name{str_firstup}
\alias{str_firstup}
\title{Sets first letter of a string to capital}
\usage{
str_firstup(x)
}
\arguments{
\item{x}{The string you want to apply the formatting to}
}
\value{
A formatted string
}
\description{
Sets first letter of a string to capital
}
\examples{
str_firstup("ThIs iS One uGLy sTRIng!")
}
