% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_tools.R
\name{str_mid}
\alias{str_mid}
\title{Returns a string containing a specified number of characters from a string}
\usage{
str_mid(text, idx_start, qty_characters = NA)
}
\arguments{
\item{text}{The string you want the get a piece of}

\item{idx_start}{Character position in string at which the part to be taken begins. If start is greater than the number of characters in string, NA will be returned.}

\item{qty_characters}{If omitted or if there are fewer than length characters in the text (including the character at start), all characters from the start position to the end of the string are returned.}
}
\value{
A string
}
\description{
Returns a string containing a specified number of characters from a string
}
\examples{
str_mid(text = "A long, long time ago", idx_start = 89)
str_mid(text = "A long, long time ago", idx_start = 9, qty_characters = 9)
}
