% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_tools.R
\name{str_right}
\alias{str_right}
\title{Extracts a substring from a string, starting from the right-most character}
\usage{
str_right(text, qty_characters)
}
\arguments{
\item{text}{The string you want the get a piece of}

\item{qty_characters}{If  the number of characters that you wish to extract starting from the left-most character.}
}
\value{
A string
}
\description{
Extracts a substring from a string, starting from the right-most character
}
\examples{
str_right(text = "A long, long time ago", qty_characters = 9)
}
