% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tbl_companies_uk}
\alias{tbl_companies_uk}
\title{A sample of 1.835 UK companies}
\format{A data frame with 1.835 rows and 23 variables:
\describe{
  \item{GDB_ORG_YUID}{Company identifier}
  \item{GB_CRO_REGISTRATION_NUMBER}{Chamber of commerce identifier}
  \item{GB_Y_COMPANY_ID}{Legacy company identifier}
  \item{PRINCIPAL_NAME}{Legacy company identifier}
  \item{ADDRESS_LINE_1_RA}{Address line for the registred address}
  \item{ADDRESS_LINE_2_RA}{Address line for the registred address}
  \item{ADDRESS_LINE_3_RA}{Address line for the registred address}
  \item{ADDRESS_LINE_4_RA}{Address line for the registred address}
  \item{ADDRESS_LINE_5_RA}{Address line for the registred address}
  \item{POSTCODE_RA}{Legacy company identifier}
  \item{POSTCODE_RA}{Legacy company identifier}
}}
\usage{
tbl_companies_uk
}
\description{
A dataset containing the first few companies from our UK GDI dataset
}
\keyword{datasets}
