% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tbl_company_relations}
\alias{tbl_company_relations}
\title{Relations between companies}
\format{A data frame with 392.835 rows and 2 variables:
\describe{
  \item{id_company}{Code used to identify economic activities}
  \item{id_company_parent}{Code of the economic activity description that is more general}
}}
\usage{
tbl_company_relations
}
\description{
Relations between companies
}
\keyword{datasets}
