% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tbl_NACE}
\alias{tbl_NACE}
\title{The 2008 NACE codes which which are used to describe a companies main economic activity.}
\format{A data frame with 1.455 rows and 4 variables:
\describe{
  \item{code_NACE}{Code used to identify economic activities}
  \item{code_NACE_parent}{Code of the economic activity description that is more general}
  \item{description_NACE_nl}{The Dutch description of the economic activity}
  \item{description_NACE_fr}{The French description of the economic activity}
  \item{description_NACE_de}{The German description of the economic activity}
  \item{description_NACE_en}{The English description of the economic activity}
  \item{hierarchy layer}{The hierarchical position of the code in the total economic activity tree;
  the higher the number, the more specific the code is.}
}}
\usage{
tbl_NACE
}
\description{
\url{https://ec.europa.eu/eurostat/documents/3859598/5902521/KS-RA-07-015-EN.PDF}
}
\keyword{datasets}
