% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tbl_postcode_4_nl}
\alias{tbl_postcode_4_nl}
\title{The postcode table for The Netherlands}
\format{A data frame with 4.770 rows and 7 variables:
\describe{
  \item{postcode_4}{4 digit postcode}
  \item{town}{Town}
  \item{municipality}{Municipality}
  \item{province}{Province}
  \item{lat}{Latitude}
  \item{lon}{Longitude}
  \item{type_of_address}{Type of postcode use}
}}
\usage{
tbl_postcode_4_nl
}
\description{
\url{https://ec.europa.eu/eurostat/documents/3859598/5902521/KS-RA-07-015-EN.PDF}
}
\keyword{datasets}
