% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tbl_SBI}
\alias{tbl_SBI}
\title{The SBI codes which which are used to describe a companies main economic activity.}
\format{A data frame with 1.887 rows and 4 variables (previously 1.455 rows):
\describe{
  \item{code_SBI}{Code used to identify economic activities}
  \item{code_SBI_parent}{Code of the economic activity description that is more general}
  \item{description_SBI}{The Dutch description of the economic activity}
  \item{hierarchy layer}{The hierarchical position of the code in the total economic activity tree;
  the higher the number, the more specific the code is.}
}}
\usage{
tbl_SBI
}
\description{
A dataset correcting character problem of original source
}
\details{
\url{https://sbi.cbs.nl/cbs.typeermodule.typeerservicewebapi/content/angular/app/#/}
}
\keyword{datasets}
