% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tbl_SBI_count}
\alias{tbl_SBI_count}
\title{The SBI codes which which are used to describe a companies main economic activity and the number of companies in nl}
\format{A data frame with 1.455 rows and 4 variables:
\describe{
  \item{code_SBI}{Code used to identify economic activities}
  \item{code_SBI_parent}{Code of the economic activity description that is more general}
  \item{description_SBI}{The Dutch description of the economic activity}
  \item{hierarchy_layer}{The hierarchical position of the code in the total economic activity tree;
  the higher the number, the more specific the code is.}
  \item{qty_companies}{The number of companies in The Netherlands}
}}
\usage{
tbl_SBI_count
}
\description{
\url{https://sbi.cbs.nl/cbs.typeermodule.typeerservicewebapi/content/angular/app/#/}
}
\keyword{datasets}
