% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tbl_SBI_new}
\alias{tbl_SBI_new}
\title{DO NOT USE THIS DATA FRAME. PLEASE USE tbl_SBI}
\format{An object of class \code{data.frame} with 1887 rows and 4 columns.}
\usage{
tbl_SBI_new
}
\description{
This data frame is used as a backup.
}
\keyword{datasets}
