% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_tools.R
\name{theme_graydon}
\alias{theme_graydon}
\title{#' A vector with the Graydon color palette
#' @export
col_graydon <- c("#00545D", "#EB6E08", "#858587",
                 "#8BB0DE", "#186FA7", "#474646",
                 "#BABFC1", "#000000", "#3AA2DF")
#' Color from the Graydon palette representing a low value
#' @export
col_graydon_low <- col_graydon[4]
#' Color from the Graydon palette representing a high value
#' @export
col_graydon_high <- col_graydon[2]
#' Color from the Graydon palette for the grid
#' @export
col_graydon_grid <- col_graydon[7]
#' Color from the Graydon palette for the axes
#' @export
col_graydon_axis <- col_graydon[6]
devtools::use_data(col_graydon_axis)
A function for applying a Graydon theme to a ggplot}
\usage{
theme_graydon(type = c("grid", "horizontal", "vertical", "blank"))
}
\arguments{
\item{type}{The type of grid or background}
}
\description{
This function applies Graydon theme to a ggplot.
The first time it is used it will create a column specification file
allows you to create and or read a file which you can use to control a data file import
}
\examples{
theme_graydon("grid")
}
\keyword{ggplot2}
