% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/company_hierarchy.R
\name{total_hierarchy_value}
\alias{total_hierarchy_value}
\title{Adds an total value to the vertices of a company hierarchy}
\usage{
total_hierarchy_value(graph, name_attribute, name_total, FUN, ...)
}
\arguments{
\item{graph}{A graph}

\item{name_attribute}{The name of the value attribute to be aggregated}

\item{name_total}{The name of the attribute where the total value is stored}

\item{FUN}{the function which is used to calculate total value}

\item{...}{The parameters passed to the function specified in FUN}
}
\value{
A graph where all the nodes contain aggregated value
}
\description{
Adds an total value to the vertices of a company hierarchy
}
\examples{
graph <- total_hierarchy_value(graph, name_attribute = "qty_employees", name_total = "qty_employees_sum", FUN = sum, na.rm = TRUE)
}
\keyword{company}
\keyword{graph}
\keyword{hierarchy}
