% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/economic_activity_hierachy.R
\name{vertices_add_distance_to_root}
\alias{vertices_add_distance_to_root}
\title{Add the distance to each vertex to the graph's root}
\usage{
vertices_add_distance_to_root(graph, vertex_attribute = "dist_to_root")
}
\arguments{
\item{graph}{The graph to do the searhing in}

\item{vertex_attribute}{the name you want the vertex distance attribute to have}
}
\value{
Graph
}
\description{
Add the distance to each vertex to the graph's root
}
\keyword{NACE}
\keyword{SBI}
\keyword{SIC}
