% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_config.R
\name{write_colspecs_file}
\alias{write_colspecs_file}
\title{Write column specs to file}
\usage{
write_colspecs_file(tbl_col_specs, filename_columns)
}
\arguments{
\item{tbl_col_specs}{A data-frame containing the column specifications}

\item{filename_columns}{The file to write the column specifications to}
}
\description{
Write column specs to file
}
\examples{
tbl_column_specs <- read_colspecs_file(filename_specs)
}
\keyword{config}
\keyword{data}
\keyword{import}
