% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_tools.R
\docType{data}
\name{col_graydon}
\alias{col_graydon}
\title{A vector with the Graydon color palette}
\format{An object of class \code{character} of length 9.}
\usage{
col_graydon
}
\description{
A vector with the Graydon color palette
}
\keyword{datasets}
