% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/economic_activity_hierachy.R
\name{create_economic_activity_graph}
\alias{create_economic_activity_graph}
\title{Making a graph of a economic activity hierarchy data-frame.}
\usage{
create_economic_activity_graph(tbl_hierarchy, col_id = "code",
  col_id_parent = "code_parent")
}
\arguments{
\item{tbl_hierarchy}{The data frame containing all codes and references to their parents.}

\item{col_id}{The name of the column that is the economic activity code}

\item{col_id_parent}{The name of the column that is the parent's code of the economic activity code}
}
\value{
Graph representation of the economic activity hierarchy
}
\description{
Making a graph of a economic activity hierarchy data-frame.
}
\examples{
graph_SBI <- create_economic_activity_graph(tbl_SBI_count, col_id = "code_SBI", col_id_parent = "code_SBI_parent")
}
\keyword{NACE}
\keyword{SBI}
\keyword{SIC}
