% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/economic_activity_hierachy.R
\name{get_root_vertex}
\alias{get_root_vertex}
\title{Get the root node of a tree}
\usage{
get_root_vertex(tree_graph)
}
\arguments{
\item{tree_graph}{The graph containing the hierarchical tree}
}
\value{
The vertext that is the root of the graph
}
\description{
Get the root node of a tree
}
\examples{
graph_SBI <- create_economic_activity_graph(tbl_SBI_count, col_id = "code_SBI", col_id_parent = "code_SBI_parent")
}
\keyword{NACE}
\keyword{SBI}
\keyword{SIC}
