% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uncategorized_tools.R
\name{open_project}
\alias{open_project}
\title{Creates a subdirectory and/or sets the working directory of a project. ----}
\usage{
open_project(project_name, dir_base = NULL)
}
\arguments{
\item{project_name}{The name of the subdirectory for the project you use/create}

\item{dir_base}{The directory where you want to have the project created in. The default is the current working directory}
}
\description{
Creates a subdirectory and/or sets the working directory of a project. ----
}
\examples{
open_project("My project")
open_project("My project" , "~/R Scripts")
}
