% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/economic_activity_hierachy.R
\name{roll_up_hierarchy_by_minimum}
\alias{roll_up_hierarchy_by_minimum}
\title{Roll up economic activity hierarchy so that the codes hold a minimum value}
\usage{
roll_up_hierarchy_by_minimum(graph_tree, name_attribute, name_propagated,
  threshold)
}
\arguments{
\item{graph_tree}{Graph representing the economic activity hierarchy}

\item{name_attribute}{The name of the attribute you want to base the roll-up on}

\item{name_propagated}{The name of the new attribute that should contain the cumulative value of name_attribute}

\item{threshold}{The minimum value name_attribute should have to be exluded from further roll-up}
}
\value{
A data frame containing the original economic activity code, the new activity code and the quantity/value that the new code would contain if aggregated
}
\description{
This function can be used to aggregate a NACE or SBI code economic activity tree,
so the codes represent enough of something for example number of companies, number of customers or
revenue.
}
\keyword{NACE}
\keyword{SBI}
\keyword{SIC}
