% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/company_hierarchy.R
\name{select_graph_hierarchies}
\alias{select_graph_hierarchies}
\title{Function to find the hierarchies of}
\usage{
select_graph_hierarchies(graph_company_hierarchies, id_companies)
}
\arguments{
\item{graph_all_companies}{A graph containing all company/company relations data.}

\item{id_company}{The id of the company of which you want to retrieve the whole hierarchy}
}
\value{
A graph with the company hierarchy of the specifief company
}
\description{
This function searches the graph of a company's complete hierarchy from a graph containing
a multitude of company hierarchies.
}
\examples{
graph_company_hierarchy <- find_company_hierarchy(graph_company_hierarchies, "931238099")
}
\keyword{company}
\keyword{graph}
\keyword{hierarchy}
