% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_tools.R
\name{theme_graydon}
\alias{theme_graydon}
\title{A function for applying a Graydon theme to a ggplot}
\usage{
theme_graydon(type = c("grid", "horizontal", "vertical", "blank"))
}
\arguments{
\item{type}{The type of grid or background}
}
\description{
This function applies Graydon theme to a ggplot.
The first time it is used it will create a column specification file
allows you to create and or read a file which you can use to control a data file import
}
\examples{
get_columns_file(df, filename_source)
}
\keyword{ggplot2}
